+++ rooble/Gemfile
@@ -1,0 +1,9 @@
+source "https://rubygems.org"
+
+gem 'rspec'
+
+group :development do
+    gem 'guard'
+    gem 'guard-rspec'
+    gem 'pry'
+end
